requirements.txt
streamlit
torch
diffusers
transformers
Pillow
requests
auth-token